:fire: Remove gradio and use streamlit
