--- aide/webui/theme.py
@@ -1,75 +1,0 @@
-from __future__ import annotations
-from typing import Iterable
-import gradio as gr
-from gradio.themes.base import Base
-from gradio.themes.utils import colors, fonts, sizes
-import time
-
-class WecoAI(Base):
-    def __init__(
-        self,
-        *,
-        primary_hue: colors.Color | str = colors.pink,
-        secondary_hue: colors.Color | str = colors.fuchsia,
-        neutral_hue: colors.Color | str = colors.gray,
-        spacing_size: sizes.Size | str = sizes.spacing_md,
-        radius_size: sizes.Size | str = sizes.radius_md,
-        text_size: sizes.Size | str = sizes.text_lg,
-        font: fonts.Font
-        | str
-        | Iterable[fonts.Font | str] = (
-            fonts.GoogleFont("Quicksand"),
-            "ui-sans-serif",
-            "sans-serif",
-        ),
-        font_mono: fonts.Font
-        | str
-        | Iterable[fonts.Font | str] = (
-            fonts.GoogleFont("IBM Plex Mono"),
-            "ui-monospace",
-            "monospace",
-        ),
-    ):
-        super().__init__(
-            primary_hue=primary_hue,
-            secondary_hue=secondary_hue,
-            neutral_hue=neutral_hue,
-            spacing_size=spacing_size,
-            radius_size=radius_size,
-            text_size=text_size,
-            font=font,
-            font_mono=font_mono,
-        )
-        super().set(
-            body_background_fill="#F2F0E7",
-            body_background_fill_dark="repeating-linear-gradient(45deg, *primary_800, *primary_800 10px, *primary_900 10px, *primary_900 20px)",
-            button_primary_background_fill="linear-gradient(90deg, *primary_300, *secondary_400)",
-            button_primary_background_fill_hover="linear-gradient(90deg, *primary_200, *secondary_300)",
-            button_primary_text_color="white",
-            button_primary_background_fill_dark="linear-gradient(90deg, *primary_600, *secondary_800)",
-            slider_color="*secondary_300",
-            slider_color_dark="*secondary_600",
-            block_title_text_weight="600",
-            block_border_width="3px",
-            block_shadow="*shadow_drop_lg",
-            button_primary_shadow="*shadow_drop_lg",
-            button_large_padding="32px",
-        )
-
-wecoai = WecoAI()
-
-with gr.Blocks(theme=wecoai) as demo:
-    textbox = gr.Textbox(label="Name")
-    slider = gr.Slider(label="Count", minimum=0, maximum=100, step=1)
-    with gr.Row():
-        button = gr.Button("Submit", variant="primary")
-        clear = gr.Button("Clear")
-    output = gr.Textbox(label="Output")
-
-    def repeat(name, count):
-        time.sleep(3)
-        return name * count
-
-    button.click(repeat, [textbox, slider], output)
-
-demo.launch()
